requirements.txt
# Core dependencies
langchain>=0.1.0
langchain-community>=0.0.13
langchain-openai>=0.0.5
openai>=1.12.0

# PDF processing
pypdf>=4.0.0
pdfplumber>=0.10.0

# Vector database
faiss-cpu>=1.7.4

# Embeddings
sentence-transformers>=2.3.0

# Evaluation and metrics
ragas>=0.1.0
rouge-score>=0.1.2
nltk>=3.8.0

# Utilities
python-dotenv>=1.0.0
tiktoken>=0.5.0
tqdm>=4.66.0
pandas>=2.0.0
numpy>=1.24.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0

# Streamlit for demo UI (optional)
streamlit>=1.31.0
